SCRUM-61 Sabotageable Torches (Client-side)
Torches can now be sabotaged when your infected, however it's only for the infected player at the moment. The server needs to replicate the torch rather than the client creating it themselves from the seed.

As an infected player, I want to be able to sabotage the lights so that I can infect players easier.